auto_checker/requirements.txt
django>=4.2,<5.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
PyPDF2>=3.0.0
